pyproject.toml
[tool.poetry]
name = "dashboards"
version = "1.0.0"
description = "Personal dashboard collection - Auto-updating static sites"
authors = ["Austin Wagner"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.5"
playwright = "^1.57.0"
pyyaml = ">=6.0"
jinja2 = ">=3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
